requirements.txt
flask==3.0.2
pandas==1.5.3
gunicorn==20.1.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
numpy==1.24.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0

# API Integration
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.0.3
aiohttp==3.8.5
cachetools==5.3.1

# Data Processing
openpyxl==3.1.2
python-magic==0.4.27

# Additional dependencies
psutil==5.9.8
redis==5.0.1
sqlalchemy==2.0.27
pytest==7.4.3
pytest-cov==4.1.0

# Development dependencies
certifi==2024.2.2
prometheus-client==0.19.0
structlog==24.1.0
tenacity==8.2.3
typing-extensions==4.9.0
urllib3==2.2.0
pytest-asyncio==0.23.5
black==24.1.1
isort==5.13.2
mypy==1.7.1
flake8==7.0.0
pylint==3.0.3
